Created Footer & About components & updated Realtor table
Added social media fields for the realtor table in the database through the Strapi admin page so Brian can add/edit/delete his social media account info.  Completed the Footer component structure (still have to bring in the getBrian action creator to tie the links to this account info) and completed the navbar for mobile view.  Desktop navbar is going to have some react animations on scroll so that will be a continuing work-in-progress.  Started the "About Brian" section by creating the getBrian action creator and updating the reducer.  Lastly, added the fontawesome cdn to index.html.