Footer & Navbar Improvements
Brought in the getBrianInfo (that I incorrectly labeled as getBrian in the last commit) action creator into the footer and fetched Brian's info so all of the links now point to his socials and real contact information.  I still have to add in conditionals that check to make sure the value is not null or empty - if they are, we're not going to render the icon.  Added "sticky-top" to the .collapse section of the navbar so it will appear on the screen no matter where the end user clicks the toggle button from.  Also added "top: 110px" so it would not overlap the navbar header on scroll.