Added Action Creators & New Components
All photos fall under the '/upload' endpoint so I had to make another action creator to retrieve Brian's headshot - added the type to types.js and included it in the realtorReducer.  Added a spinner.gif and Spinner component so I can do conditionals in the jsx to check for loading. Uninstalled some npm packages I'm not going to use. Added the remaining components that will be featured on the Home page so I could visualize the spacing.  Currently working on converting markdown to jsx as the text that comes from the db through Strapi comes as markdown.