Converted markdown text to jsx & background color/image work
npm installed markdown-to-jsx so any formatting that Brian does in the WYSIWYG text editor, in Strapi admin for his "About" section, will show as is.  Had to use the render method and setTimeout function to make it work.  Created the Testimonial component and used background: linear-gradient and adobe xd to make the website a lot less blue/grey.