testimonial text fix and started recent listing component
Noticed when I entered a rather large paragraph of text for a testimonial it overflowed the container (because I have a set 40vh for the div.row).  Tried multiple solutions to try to get the text to auto re-size with no luck.  Ended up applying font-size: 85% to the carousel-item which contains the testimonials we currently have.  I really don't want to turn off the 40vh, which would go back to the default of  height: auto for the container-fluid - it would fit the text at bigger font-sizes but the div itself would constantly change in size every transition. Will likely need a better solution long term.