Created two more components & more css work
Put together the layout for the MortgageCalculator component - still have to put together the API call. Created Listing component and AllListings component.  I re-used the Listing componenet in the RecentListing component and will re-use it in the AllListings component to map through all listings in the database.  Still have to hook up action creators and reducers.