Set up axios GET request to Zillow API
Was able to successfully send a GET request to Zillow API for Current Mortgage Rates.  For now I am storing my 'partnerId' in an environment variable but I have read numerous times this is bad practice as it will still be exposed in the browser if someone wants to find it in the files via inspect element.  Strapi has seven built-in 'Providers' that you can make a call to, none of which is Zillow.  I might have to create a second backend server to handle the API calls to Zillow in order to protect my partnerId. I'm still trying to figure out if the npm config package will suffice for security.