Created new components and the listingReducer
Created the RecentListing, AllListings, ListingDetail, and Listings components, along with the listingReducer.  I decided to restructure how I would render each listing.  It made sense to separate the rendering of the most recent listing into it's own component as it will only be shown on the home page. I also moved the navbar into App.js as it will be the same for every page on mobile view, the only variation will be in desktop view, so I will handle those cases by page component most likely.

Added conditionals to the footer that only render contact info  and icons when the data is present in the database.  If Brian decided to delete any field, it will disappear and the container-fluid will reformat the content.