Created listing detail route & minor updates
During my last commit I also created the getListing function to retrieve the details of a given listing when "more info" is clicked. I set up the route for the ListingDetail component and began creating the UI for it.  Mistakenly initialized allListings with an empty object instead of an empty array, so I switched that to make mapping a bit easier. Added some zoom animation to the cover photos and thumbnail images of listings. Early on I was figuring out how I wanted to do naming conventions and decided on using the exact name of the reducer when mapping state to props, therefore, I switched testimonials_object to testimonialsReducer.