Added fields to Listings table in Strapi & component work
Brian also does rental properties so I had to account for this with 'Property Type' (e.g. residential, apartment, condo, etc.) and 'Availability' (e.g. for sale, for rent) fields in the Listings table.
Just about done with the ListingDetail component and I have the PicModal component up and working.  Still have to pass in the corresponding photo url onClick and finish general css/design of that component.