Completed PicModal component
I wanted to show one main, large photo with a carousel of thumbnail images underneath for the end user to choose from - I npm i react-multi-carousel to accomplish this.  This package will also come in handy when I want to display 2 or 3 testimonials on larger screens.  I am going to implement the same feature on the main ListingDetail component so all of the photos for a given property are not shown - there will be the cover photo + 2-3 on mobile and several on larger screens.