Added Carousel to ListDetail component & back buttons
On mobile view only 2 photos show in the carousel under the main cover photo.  The forward and back arrows will take you through the pictures but onClick of any photo will still generate the PicModal component.  I added a < symbol at the top of ListDetail and a "Back to Listings" button at the bottom of the page to make it easy for people to navigate back to browsing.