Changed back button behavior and inner text
My original intent was to use both back buttons in the ListDetail to send someone back exactly where they came from.  I needed to use history.goBack(), not .push(), so I changed that.  I also changed the inner text on the "Back to Listings" to "Go Back". If someone clicks on "More Info" for the Most Recent Listing on the Home page and clicks back, it will take them back to the Home page, not the Listings page, so semantically "Back to Listings" was incorrect.