Created Inquiry & Contact tables in the database
Tomorrow I will be putting together the inquiry and contact forms so I set up the tables and fields for the POST requests. I want the inquiry form to be a modal that will only be accessible from the ListingDetail component.  I also have an inquiry button in the PicModal component but I'm unsure if nested modals is possible, nor am I sure if it's good practice or design. My initial idea is to set showPicModal to false onClick, then render the InquiryModal component.