Worked on components & altered some table fields in the db
I added a 'description' field to the landing table as we need it for the alt attribute on the <img> tag.  Eliminated the '?' from 'published' on tables with 'published' as it was being sent as part of the data to the frontend. Put together the InquiryForm.  Tomorrow I will put together the reducer and action for it.