Updated contact form & minor fixes
I realized I forgot to add in the regex to make sure the phone number input is 10 digits, not just 10 characters of any kind.

For the contact form, as of now the only valid input that could be sent and rejected on the server side is if the question is identical to another question in the database. I need to tie this to an email just on the off chance someone actually does word their question the same as someone else.
I would also like to give more meaningful error messages on the client side depending on the mongo error, but until I figure that out I wrote a conditional in the Alert component checking for the message it would receive otherwise. This is problematic as an internal server error does not necessarily have to be what I assigned it to be ('Your message appears to have already been sent.')

I'm trying to implement Sentry as my application monitoring software, but following Strapi's documentation on it hasn't led me to success. I will continue to figure this out.