Worked on login, registration, logout & last component
The User table is all setup and the login, registration and logout work, however, I forgot to setup the OAuth for Google to give the users a second option for connecting, so I will take care of that tomorrow.  The news API I was originally going to use for the news and events page is not as good as I thought.  The results are almost unrelated the query params.  I might use Bing or New York Times. We'll see.