Created table for google authentication users
Since the Users table on Strapi requires a "username" and it appears I am unable to edit the fields of that table, I created a table for users who register via Google OAuth.  It only records the user's first name, last name, and email.  I've been working on the design for larger screen breakpoints but it's ready for deployment and I can finish designing for breakpoints live.  I'm going to have the site hidden until I'm done with the design.